Function teaclave_client_sdk::bindings::teaclave_register_output_file_serialized

#[no_mangle]
pub unsafe extern "C" fn teaclave_register_output_file_serialized(
    client: &mut FrontendClient,
    serialized_request: *const c_char,
    serialized_response: *mut c_char,
    serialized_response_len: *mut size_t
) -> c_int

Send JSON serialized request to the service with the client and get the serialized response.

Arguments

client: service client.

serialized_request; JSON serialized request

serialized_response: buffer to store the JSON serialized response.

serialized_response_len: length of the allocated serialized_response, will be set as the length of serialized_response when return successfully.

Return

The function returns 0 for success. On error, the function returns 1.

Safety

Inconsistent length of allocated buffer may caused overflow.