Module teaclave_client_sdk::bindings

Functions

teaclave_approve_task_serialized⚠
Send JSON serialized request to the service with the client and get the serialized response.

teaclave_assign_data_serialized⚠
Send JSON serialized request to the service with the client and get the serialized response.

teaclave_close_authentication_service⚠
Close and free the authentication service handle, i.e., the AuthenticaionClient type opaque pointer. The function returns 0 for success. On error, the function returns 1.

teaclave_close_frontend_service⚠
Close and free the frontend service handle, i.e., the FrontendClient type opaque pointer. The function returns 0 for success. On error, the function returns 1.

teaclave_connect_authentication_service⚠
Connect to Teaclave Authentication Service.

teaclave_connect_frontend_service⚠
Connect to Teaclave Frontend Service.

teaclave_create_task_serialized⚠
Send JSON serialized request to the service with the client and get the serialized response.

teaclave_get_function_serialized⚠
Send JSON serialized request to the service with the client and get the serialized response.

teaclave_get_task_result⚠
Get task result of task_id. The result will be save in the task_result buffer, and set corresponding task_result_len argument. Note that this is a blocking function and wait for the return of the task. The function returns 0 for success. On error, the function returns 1.

teaclave_get_task_serialized⚠
Send JSON serialized request to the service with the client and get the serialized response.

teaclave_invoke_task⚠
Invoke task with task_id. The function returns 0 for success. On error, the function returns 1.

teaclave_invoke_task_serialized⚠
Send JSON serialized request to the service with the client and get the serialized response.

teaclave_register_function_serialized⚠
Send JSON serialized request to the service with the client and get the serialized response.

teaclave_register_input_file_serialized⚠
Send JSON serialized request to the service with the client and get the serialized response.

teaclave_register_output_file_serialized⚠
Send JSON serialized request to the service with the client and get the serialized response.

teaclave_set_credential⚠
Set user's credential with user_id and user_token. The function returns 0 for success. On error, the function returns 1.

teaclave_user_login⚠
Login a new user with user_id and user_password. The login session token will be save in the token buffer, and length will be set in the token_len argument. The function returns 0 for success. On error, the function returns 1.

teaclave_user_login_serialized⚠
Send JSON serialized request to the service with the client and get the serialized response.

teaclave_user_register⚠
Register a new user with user_id and user_password. The function returns 0 for success. On error, the function returns 1.

teaclave_user_register_serialized⚠
Send JSON serialized request to the service with the client and get the serialized response.